Add `@github/markdown-toolbar-element` dependency
The ActionMarkdown toolbar will rely on the
`@github/markdown-toolbar-element` JavaScript library. To make it easy
to install, we will install this JavaScript dependency when running the
`rails action_markdown:install` installation script.

In the `test/dummy` application, we ran the installation script with
importmap installed to pin the JavaScript dependency.